# You-book Project/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "유북 API 명세서",
    "description": "Node.js Swaager swagger-jsdoc 방식 RestFul API 클라이언트 UI"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/check_id": {
      "post": {
        "summary": "아이디 중복 확인",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "확인할 사용자 아이디"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "아이디 중복 확인 결과",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "요청 성공 여부"
                    },
                    "message": {
                      "type": "string",
                      "description": "중복 확인 메시지"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "아이디 미입력"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/login_process": {
      "post": {
        "summary": "사용자 로그인",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "사용자 아이디"
                  },
                  "pw": {
                    "type": "string",
                    "description": "사용자 비밀번호"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공 여부 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "로그인 성공 여부"
                    },
                    "message": {
                      "type": "string",
                      "description": "로그인 메시지"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "summary": "사용자 로그아웃",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "홈 화면으로 리다이렉트"
          }
        }
      },
      "post": {
        "summary": "사용자 로그아웃",
        "tags": [
          "Auth"
        ],
        "description": "세션을 파기하고 로그아웃 처리. 로그아웃 후 쿠키 삭제.",
        "responses": {
          "200": {
            "description": "성공적으로 로그아웃되었거나 세션이 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "로그아웃 결과 메시지"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "로그아웃 처리 중 오류 발생",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "오류 메시지"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register_process": {
      "post": {
        "summary": "사용자 회원가입",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "사용자 아이디"
                  },
                  "pw": {
                    "type": "string",
                    "description": "비밀번호"
                  },
                  "pw2": {
                    "type": "string",
                    "description": "비밀번호 확인"
                  },
                  "username": {
                    "type": "string",
                    "description": "사용자 이름"
                  },
                  "email": {
                    "type": "string",
                    "description": "이메일 주소"
                  },
                  "phone_num": {
                    "type": "string",
                    "description": "전화번호"
                  },
                  "birth": {
                    "type": "string",
                    "format": "date",
                    "description": "생일"
                  },
                  "gender": {
                    "type": "string",
                    "description": "성별"
                  },
                  "profileImage": {
                    "type": "string",
                    "format": "binary",
                    "description": "프로필 이미지 파일"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "회원가입 성공 여부 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "요청 성공 여부"
                    },
                    "message": {
                      "type": "string",
                      "description": "성공 또는 오류 메시지"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  },
...
}